fix: resolve issues #1210, #1213, #1216
Issue #1213 — Remove dead commented-out stubs
- Remove 'TODO: Uncomment when needed' aggregateFromNetwork block from
  StatsService.js (lines 621-663). Block referenced non-imported symbols
  (Horizon, Database) and was never wired up.
- Remove identical stub from LeaderboardStatsService.js (lines 373-417).
  Git history preserves the code if needed in future.

Issue #1216 — Consolidate duplicate rate-limit header helper
- buildRateLimitHeaders() was implemented twice: once in perKeyRateLimit.js
  (canonical, used by rbac.js) and once in rateLimitHeaders.js (dead module,
  never imported anywhere).
- rateLimitHeaders.js is now a thin re-export shim from perKeyRateLimit.js,
  maintaining backward compatibility while eliminating the duplicate.
- Add docs/RATE_LIMIT_COMPOSITION.md describing the three rate-limit layers,
  their responsibilities, and composition order.

Issue #1210 — Decompose 2813-line donation route module
- Split src/routes/donation.js into four focused sub-routers under
  src/routes/donations/:
    create.js  — POST /send, POST /, POST /batch, POST /bulk,
                 POST /cross-asset, POST /claimable, POST /claimable/:id/claim
    export.js  — POST /export, GET /export/:jobId,
                 GET /export/:jobId/download, GET /export (deprecated)
    query.js   — all GET read endpoints incl. SSE status stream and stats
    notes.js   — per-donation actions: receipt, refund, tags, status, impact
    helpers.js — shared schemas, applyNotePrivacy, withDonorLock, formatBatch*
- src/routes/donation.js is now a ~50-line thin index that mounts sub-routers.
- All existing URL paths are preserved — zero breaking changes.

Closes #1210, #1213, #1216
